pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "wtforms>=3.2.1",
    "werkzeug>=3.1.3",
    "trafilatura>=2.0.0",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
    "sendgrid>=6.12.4",
    "sqlalchemy>=2.0.42",
    "pillow>=11.3.0",
    "bleach>=6.2.0",
]
